All Scripts, Except (Laser D, and Enemy H A P 1)
Basically what I have done is reenter all of the scripts except the
enemy alert pass 1 script, and the laser detection script. I haven't
made Any changes to the scene, or prefabs. This means that everything
will need to be connected again, in the editor and in the scene1.

NOTE: Everything seems to be working fine.